requirements.txt
anyio==4.4.0
certifi==2024.6.2
document==1.0
docx==0.2.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
pillow==10.3.0
python-docx==1.1.2
python-telegram-bot==21.3
sniffio==1.3.1
typing_extensions==4.12.2
